+++ work/models.py
@@ -11,7 +11,7 @@ from djangocms_text_ckeditor.fields import HTMLField
 from aldryn_translation_tools.models import TranslationHelperMixin
 from aldryn_common.admin_fields.sortedm2m import SortedM2MModelField
 
-from allink_core.allink_base.models import AllinkManualEntriesMixin, AllinkBaseModelManager
+from allink_core.allink_base.models import AllinkManualEntriesMixin, AllinkBaseModelManager, AllinkTranslatedAutoSlugifyMixin
 from allink_core.allink_base.models import AllinkBaseModel, AllinkBaseImage, AllinkBaseAppContentPlugin
 from allink_core.allink_base.models.mixins import AllinkTranslatedAutoSlugifyMixin