+++ config/cms_toolbars.py
@@ -7,22 +7,21 @@ from cms.cms_toolbars import ADMIN_MENU_IDENTIFIER, ADMIN_SITES_BREAK
 from cms.toolbar_pool import toolbar_pool
 from cms.toolbar_base import CMSToolbar
 
-from allink_apps.members.models import Members
+from allink_apps.config.models import Config
 
 
 @toolbar_pool.register
-class MembersToolbar(CMSToolbar):
-    model = Members
-    app_label = Members._meta.app_label
+class ConfigToolbar(CMSToolbar):
+    model = Config
 
     def populate(self):
         admin_menu = self.toolbar.get_menu(ADMIN_MENU_IDENTIFIER)
         position = admin_menu.find_first(Break, identifier=ADMIN_SITES_BREAK)
-        allink_menu = admin_menu.get_or_create_menu(
-            'allink-menu',
-            _('Allink'),
+        project_menu = admin_menu.get_or_create_menu(
+            'project-menu',
+            _('Project'),
             position=position
         )
 
         url = reverse('admin:{}_{}_changelist'.format(self.model._meta.app_label, self.model._meta.model_name))
-        allink_menu.add_sideframe_item(self.model.get_verbose_name_plural(), url=url)
+        project_menu.add_sideframe_item(_(u'Config'), url=url)