+++ contact/urls.py
@@ -1,6 +1,6 @@
 # # -*- coding: utf-8 -*-
 from django.conf.urls import url
-from .views import ContactRequestView
+from allink_apps.contact.views import ContactRequestView
 
 
 urlpatterns = [