+++ contact/views.py
@@ -4,11 +4,12 @@ from __future__ import unicode_literals
 from allink_core.allink_base.views import AllinkBaseCreateView
 from allink_core.allink_mandrill.config import MandrillConfig
 
-from .models import ContactRequest
-from .forms import ContactRequestForm
-from .email import send_request_confirmation_email, send_request_email
+from allink_apps.contact.models import ContactRequest
+from allink_apps.contact.forms import ContactRequestForm
+from allink_apps.contact.email import send_request_confirmation_email, send_request_email
 
 config = MandrillConfig()
+
 
 class ContactRequestView(AllinkBaseCreateView):
     model = ContactRequest