+++ work/cms_apps.py
@@ -11,4 +11,5 @@ class WorkApphook(CMSApp):
     def get_urls(self, page=None, language=None, **kwargs):
         return ['allink_apps.work.urls']
 
+
 apphook_pool.register(WorkApphook)