+++ work/cms_menus.py
@@ -3,12 +3,13 @@ from menus.menu_pool import menu_pool
 from django.utils.translation import ugettext_lazy as _
 from cms.menu_bases import CMSAttachMenu
 
-from .models import Work
+from allink_apps.work.models import Work
 
 
 class WorkMenu(CMSAttachMenu):
 
     name = _("Work menu")
+
     def get_nodes(self, request):
 
         nodes = []
@@ -23,5 +24,6 @@ class WorkMenu(CMSAttachMenu):
             nodes.append(node)
 
         return nodes
+
 
 menu_pool.register_menu(WorkMenu)