+++ work/sitemaps/sitemaps.py
@@ -2,7 +2,7 @@
 
 from __future__ import unicode_literals
 from django.contrib.sitemaps import Sitemap
-from ..models import Work
+from allink_apps.work.models import Work
 
 
 class WorkSitemap(Sitemap):